requirements.txt
absl-py==2.1.0
accimage @ file:///home/conda/feedstock_root/build_artifacts/accimage_1694355651270/work
ADEval==1.1.0
aiohttp==3.9.5
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.4.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
contourpy==1.1.1
cycler==0.12.1
einops==0.8.0
et-xmlfile==1.1.0
faiss-gpu==1.7.2
fastprogress==1.0.3
filelock==3.13.1
fonttools==4.53.1
FrEIA==0.2
frozenlist==1.4.1
fsspec==2024.2.0
ftfy==6.2.0
future==1.0.0
fvcore==0.1.5.post20221221
geomloss==0.2.6
google-auth==2.32.0
google-auth-oauthlib==1.0.0
grpcio==1.65.1
huggingface-hub==0.26.2
idna==3.4
imageio==2.34.2
imgaug==0.4.0
importlib_metadata==8.1.0
importlib_resources==6.4.0
iopath==0.1.10
Jinja2==3.1.3
joblib==1.4.2
kiwisolver==1.4.5
kornia==0.7.3
kornia_rs==0.1.5
lazy_loader==0.4
lightning-utilities==0.11.6
llvmlite==0.41.1
mamba-ssm @ file:///home/mdisk2/hulei/ADer/mamba_ssm-2.2.2%2Bcu118torch2.1cxx11abiFALSE-cp38-cp38-linux_x86_64.whl#sha256=2c2e82ac65bf4c84afda5c235740449a3566f51c3559d829cba34ef56478ceba
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mmcls==1.0.0rc6
mmdet==2.25.3
mmselfsup==1.0.0
model-index==0.1.11
modelindex==0.0.2
mpmath==1.3.0
multidict==6.0.5
networkx==3.0
ninja==1.11.1.1
numba==0.58.1
numpy==1.24.1
numpy-hilbert-curve==1.0.1
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.10.0.84
openpyxl==3.1.5
ordered-set==4.1.0
packaging==24.1
pandas==2.0.3
Pillow==8.4.0
portalocker==2.10.1
protobuf==3.20.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocotools==2.0.7
pyDeprecate==0.3.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==1.5.9
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.1
pyzorder==0.0.2
regex==2024.5.15
requests==2.28.1
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
-e git+https://gitee.com/nerin-hulei7/cad.git@0b7c7334b67568cb87ca61acef98f0861a09fc00#egg=segment_anything&subdirectory=SAM/segment-anything
shapely==2.0.5
six==1.16.0
sympy==1.12
tabulate==0.9.0
taming-transformers==0.0.1
tensorboard==2.14.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.4.0
terminaltables==3.1.10
test-tube==0.7.5
threadpoolctl==3.5.0
tifffile==2023.7.10
timm==0.8.15.dev0
tokenizers==0.19.1
torch==2.1.2+cu118
torchaudio==2.1.2+cu118
torchmetrics==1.4.0.post0
torchvision==0.16.2+cu118
tqdm==4.66.4
transformers==4.43.1
triton==2.1.0
typing_extensions==4.9.0
tzdata==2024.1
urllib3==1.26.13
wcwidth==0.2.13
Werkzeug==3.0.3
yacs==0.1.8
yarl==1.9.4
zipp==3.19.2
